ai-web-builder/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
redis==5.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
aiohttp==3.9.1
pillow==10.1.0
openai==1.3.7
anthropic==0.7.7
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
stripe==7.8.0
boto3==1.34.0
sentry-sdk[fastapi]==1.38.0
celery==5.3.4
flower==2.0.1